pyproject.toml
[tool.poetry]
name = "kp-regression"
version = "0.1.0"
description = ""
authors = ["Ismail Gadzhiev <ismailgadzhievff@gmail.com>"]
readme = "README.md"
packages = [{include = "kp_regression"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
numpy = "^1.26.4"
scikit-learn = "^1.5.0"
torch = "^2.3.1"
matplotlib = "^3.9.0"
openpyxl = "^3.1.3"
click = "^8.1.7"
seaborn = "^0.13.2"
pyyaml = "^6.0.1"
joblib = "^1.4.2"
pyarrow = "^16.1.0"
fastparquet = "^2024.5.0"
tqdm = "^4.66.4"
lightgbm = "^4.3.0"
xgboost = "^2.0.3"
catboost = "^1.2.5"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
